Get the bot structure to work on hello world
